rename namespace Tracing into Trace
