+++ src/Exporter/ZipkinExporter.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace OpenTelemetry\Exporter;
 
-use OpenTelemetry\Exporter\ExporterInterface;
-use OpenTelemetry\Tracing\Span;
+use OpenTelemetry\Trace\Span;
 
 /**
  * Class ZipkinExporter - implements the export interface for data transfer via Zipkin protocol
