+++ src/Trace/Status.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace OpenTelemetry\Tracing;
+namespace OpenTelemetry\Trace;
 
 class Status
 {
