+++ tests/unit/Tracing/Sampler/AlwaysOffSamplerTest.php
@@ -1,7 +1,7 @@
 <?php
 namespace OpenTelemetry\Tests\Unit\Tracing\Sampler;
 
-use OpenTelemetry\Tracing\Sampler\AlwaysOffSampler;
+use OpenTelemetry\Trace\Sampler\AlwaysOffSampler;
 use PHPUnit\Framework\TestCase;
 class AlwaysOffSamplerTest extends TestCase
 {
