+++ tests/unit/Tracing/Sampler/AlwaysOnSamplerTest.php
@@ -2,7 +2,7 @@
 
 namespace OpenTelemetry\Tests\Unit\Tracing\Sampler;
 
-use OpenTelemetry\Tracing\Sampler\AlwaysOnSampler;
+use OpenTelemetry\Trace\Sampler\AlwaysOnSampler;
 use PHPUnit\Framework\TestCase;
 
 class AlwaysOnTest extends TestCase
